Add files via upload
Large Language Models (LLMs) inherently carry
the biases contained in their training corpora,
which can lead to the perpetuation of societal
harm. As the impact of these foundation mod-
els grows, understanding and evaluating their bi-
ases becomes crucial to achieving fairness and
accuracy. We propose to study what LLMs know
about the world we live in through the lens of
geography. This approach is particularly pow-
erful as there is ground truth for the numerous
aspects of human life that are meaningfully pro-
jected onto geographic space such as culture, race,
language, politics, and religion. We show various
problematic geographic biases, which we define
as systemic errors in geospatial predictions. Ini-
tially, we demonstrate that LLMs are capable of
making accurate zero-shot geospatial predictions
in the form of ratings that show strong mono-
tonic correlation with ground truth (Spearman’s
ρ of up to 0.89). We then show that LLMs ex-
hibit common biases across a range of objective
and subjective topics. In particular, LLMs are
clearly biased against locations with lower so-
cioeconomic conditions (e.g. most of Africa) on a
variety of sensitive subjective topics such as attrac-
tiveness, morality, and intelligence (Spearman’s ρ
of up to 0.70). Finally, we introduce a bias score
to quantify this and find that there is significant
variation in the magnitude of bias across existing
LLMs. Code is available on the project website:
https://rohinmanvi.github.io/GeoLLM